+++ CHANGELOG.md
@@ -6,8 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.0.3.1] - unreleased
+### Added
+- slew and Slew and Center added to stellarium
+
 ### Fixed
 - fixes an error when the stellarium page is reloaded
+
+### Changed
+- rework Stellarium mount position
+- rework Stellarium selected object
+- rework slew and slewAndCenter function
 
 ## [1.0.3.0] - 2025-03-08
 ### Added